--- site/src/modules/resources/AgentLogs/useAgentLogs.test.tsx
@@ -1,115 +1,0 @@
-import { act, renderHook, waitFor } from "@testing-library/react";
-import { API } from "api/api";
-import * as APIModule from "api/api";
-import { agentLogsKey } from "api/queries/workspaces";
-import type { WorkspaceAgent, WorkspaceAgentLog } from "api/typesGenerated";
-import WS from "jest-websocket-mock";
-import { type QueryClient, QueryClientProvider } from "react-query";
-import { MockWorkspaceAgent } from "testHelpers/entities";
-import { createTestQueryClient } from "testHelpers/renderHelpers";
-import { useAgentLogs } from "./useAgentLogs";
-
-afterEach(() => {
-	WS.clean();
-});
-
-describe("useAgentLogs", () => {
-	it("should return existing logs without network calls if state is off", async () => {
-		const queryClient = createTestQueryClient();
-		queryClient.setQueryData(
-			agentLogsKey(MockWorkspaceAgent.id),
-			generateLogs(5),
-		);
-		const fetchSpy = jest.spyOn(API, "getWorkspaceAgentLogs");
-		const wsSpy = jest.spyOn(APIModule, "watchWorkspaceAgentLogs");
-		const { result } = renderUseAgentLogs(queryClient, {
-			...MockWorkspaceAgent,
-			lifecycle_state: "off",
-		});
-		await waitFor(() => {
-			expect(result.current).toHaveLength(5);
-		});
-		expect(fetchSpy).not.toHaveBeenCalled();
-		expect(wsSpy).not.toHaveBeenCalled();
-	});
-
-	it("should fetch logs when empty", async () => {
-		const queryClient = createTestQueryClient();
-		const fetchSpy = jest
-			.spyOn(API, "getWorkspaceAgentLogs")
-			.mockResolvedValueOnce(generateLogs(5));
-		jest.spyOn(APIModule, "watchWorkspaceAgentLogs");
-		const { result } = renderUseAgentLogs(queryClient, {
-			...MockWorkspaceAgent,
-			lifecycle_state: "ready",
-		});
-		await waitFor(() => {
-			expect(result.current).toHaveLength(5);
-		});
-		expect(fetchSpy).toHaveBeenCalledWith(MockWorkspaceAgent.id);
-	});
-
-	it("should fetch logs and connect to websocket", async () => {
-		const queryClient = createTestQueryClient();
-		const logs = generateLogs(5);
-		const fetchSpy = jest
-			.spyOn(API, "getWorkspaceAgentLogs")
-			.mockResolvedValueOnce(logs);
-		const wsSpy = jest.spyOn(APIModule, "watchWorkspaceAgentLogs");
-		new WS(
-			`ws://localhost/api/v2/workspaceagents/${
-				MockWorkspaceAgent.id
-			}/logs?follow&after=${logs[logs.length - 1].id}`,
-		);
-		const { result } = renderUseAgentLogs(queryClient, MockWorkspaceAgent);
-		await waitFor(() => {
-			expect(result.current).toHaveLength(5);
-		});
-		expect(fetchSpy).toHaveBeenCalledWith(MockWorkspaceAgent.id);
-		expect(wsSpy).toHaveBeenCalledWith(MockWorkspaceAgent.id, {
-			after: logs[logs.length - 1].id,
-			onMessage: expect.any(Function),
-			onError: expect.any(Function),
-		});
-	});
-
-	it("update logs from websocket messages", async () => {
-		const queryClient = createTestQueryClient();
-		const logs = generateLogs(5);
-		jest.spyOn(API, "getWorkspaceAgentLogs").mockResolvedValueOnce(logs);
-		const server = new WS(
-			`ws://localhost/api/v2/workspaceagents/${
-				MockWorkspaceAgent.id
-			}/logs?follow&after=${logs[logs.length - 1].id}`,
-		);
-		const { result } = renderUseAgentLogs(queryClient, MockWorkspaceAgent);
-		await waitFor(() => {
-			expect(result.current).toHaveLength(5);
-		});
-		await server.connected;
-		act(() => {
-			server.send(JSON.stringify(generateLogs(3)));
-		});
-		await waitFor(() => {
-			expect(result.current).toHaveLength(8);
-		});
-	});
-});
-
-function renderUseAgentLogs(queryClient: QueryClient, agent: WorkspaceAgent) {
-	return renderHook(() => useAgentLogs(agent, true), {
-		wrapper: ({ children }) => (
-			<QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-		),
-	});
-}
-
-function generateLogs(count: number): WorkspaceAgentLog[] {
-	return Array.from({ length: count }, (_, i) => ({
-		id: i,
-		created_at: new Date().toISOString(),
-		level: "info",
-		output: `Log ${i}`,
-		source_id: "",
-	}));
-}
